requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
mysql-connector-python
python-dotenv
pydantic
ruff
